move MatplotlibDefaultDPI to plots/helper.py
